Am facut restrictionarea utilizatorilor in functie de rol si am imbunatatit structura paginii de comenzi

+++ src/main/java/com/eliasdejan/digial_menu/controller/MenuItemController.java
@@ -1,9 +1,11 @@
 package com.eliasdejan.digial_menu.controller;
 
+import com.eliasdejan.digial_menu.model.CustomUserDetails;
 import com.eliasdejan.digial_menu.model.MenuItem;
 import com.eliasdejan.digial_menu.repository.MenuItemRepository;
 import com.eliasdejan.digial_menu.repository.MenuItemTypeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -26,7 +28,14 @@ public class MenuItemController {
     }
 
     @GetMapping("")
-    public String showIndex(Model model) {
+    public String showIndex(Model model, RedirectAttributes redirectAttributes) {
+
+        CustomUserDetails loggedInUser = (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to view menu items!");
+            return "redirect:/orders";
+        }
+
         model.addAttribute("message", model.asMap().get("message"));
         model.addAttribute("menuItems", menuItemRepository.findAll());
         model.addAttribute("MenuItem", new MenuItem());
@@ -37,6 +46,13 @@ public class MenuItemController {
 
     @PostMapping("/add")
     public String addMenuItem(@Valid MenuItem menuItem, BindingResult result, Model model,  RedirectAttributes redirectAttributes) {
+
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to add menu items!");
+            return "redirect:/orders";
+        }
+
         if (result.hasErrors()) {return "menu-items";}
         menuItemRepository.save(menuItem);
         redirectAttributes.addFlashAttribute("message", "Item was added!");
@@ -44,7 +60,13 @@ public class MenuItemController {
     }
 
     @GetMapping("/edit/{id}")
-    public String showUpdateForm(@PathVariable("id") long id, Model model) {
+    public String showUpdateForm(@PathVariable("id") long id, Model model, RedirectAttributes redirectAttributes) {
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to edit menu items!");
+            return "redirect:/orders";
+        }
+
         model.addAttribute("message", model.asMap().get("message"));
         MenuItem menuItem = menuItemRepository.findById((int) id)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid menu item Id:" + id));
@@ -56,8 +78,13 @@ public class MenuItemController {
     }
 
     @PostMapping("/edit/{id}")
-    public String updateMenuItem(@PathVariable("id") long id, @Valid MenuItem menuItem, BindingResult result,
-                                Model model, RedirectAttributes redirectAttributes) {
+    public String updateMenuItem(@PathVariable("id") long id, @Valid MenuItem menuItem, BindingResult result, Model model, RedirectAttributes redirectAttributes) {
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to edit menu items!");
+            return "redirect:/orders";
+        }
+
         if (result.hasErrors()) {
             menuItem.setId((int) id);
             return "/menu-items/edit/"+id;
@@ -71,6 +98,12 @@ public class MenuItemController {
 
     @GetMapping("/delete/{id}")
     public String deleteMenuItem(@PathVariable("id") long id, Model model, RedirectAttributes redirectAttributes) {
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to delete menu items!");
+            return "redirect:/orders";
+        }
+
         MenuItem menuItem = menuItemRepository.findById((int) id)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid menu item Id:" + id));
         menuItemRepository.delete(menuItem);
