+++ src/main/java/com/eliasdejan/digial_menu/controller/UserController.java
@@ -2,9 +2,11 @@ package com.eliasdejan.digial_menu.controller;
 
 import javax.validation.Valid;
 
+import com.eliasdejan.digial_menu.model.CustomUserDetails;
 import com.eliasdejan.digial_menu.model.User;
 import com.eliasdejan.digial_menu.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -26,8 +28,15 @@ public class UserController {
     }
 
     @GetMapping("")
-    public String showIndex(Model model) {
+    public String showIndex(Model model, RedirectAttributes redirectAttributes) {
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to view users!");
+            return "redirect:/orders";
+        }
+
         model.addAttribute("message", model.asMap().get("message"));
+
         // aceasta este functia care raspunde la /users/ si afiseaza o lista cu userii.
         model.addAttribute("users", userRepository.findAll()); //transmite pe frontend o lista cu toti userii prin variabila users
         model.addAttribute("User", new User()); //transmite pe frontend un user nou in caz ca se doreste crearea unui nou user
@@ -36,8 +45,15 @@ public class UserController {
 
     @PostMapping("/add")
     public String addUser(@Valid User user, BindingResult result, Model model,  RedirectAttributes redirectAttributes) {
+
         if (result.hasErrors()) {
             return "users/index";
+        }
+
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to create users!");
+            return "redirect:/orders";
         }
 
         userRepository.save(user);
@@ -46,9 +62,15 @@ public class UserController {
     }
 
     @GetMapping("/edit/{id}")
-    public String showUpdateForm(@PathVariable("id") long id, Model model) {
+    public String showUpdateForm(@PathVariable("id") long id, Model model, RedirectAttributes redirectAttributes) {
         User user = userRepository.findById((int) id)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
+
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin() && loggedInUser.getUserId() != user.getId()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to edit users!");
+            return "redirect:/orders";
+        }
 
         model.addAttribute("user", user);
         model.addAttribute("message", model.asMap().get("message"));
@@ -61,6 +83,12 @@ public class UserController {
         if (result.hasErrors()) {
             user.setId((int) id);
             return "/users/edit/"+id;
+        }
+
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin() && (loggedInUser.getUserId() != user.getId() || user.getIsAdmin())){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to edit users!");
+            return "redirect:/orders";
         }
 
         User oldUser = userRepository.findById((int) id)
@@ -79,6 +107,13 @@ public class UserController {
     public String deleteUser(@PathVariable("id") long id, Model model, RedirectAttributes redirectAttributes) {
         User user = userRepository.findById((int) id)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
+
+        CustomUserDetails loggedInUser = (CustomUserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!loggedInUser.getIsAdmin()){
+            redirectAttributes.addFlashAttribute("message", "You are not allowed to delete users!");
+            return "redirect:/orders";
+        }
+
         userRepository.delete(user);
         redirectAttributes.addFlashAttribute("message", "Item was deleted!");
         return "redirect:/users";
